Generalize grid map defines between core and extractors
https://github.com/cmangos/mangos-classic/commit/c3ea0498239e86a6f310458db86ef77090e01eab#diff-87173431a9bc65153788196b58044541
